refactor: 객체 지향 원칙을 지키도록 리팩토링
