+++ app/src/main/java/woowacourse/omok/model/rule/LoadMap.kt
@@ -1,5 +1,8 @@
 package woowacourse.omok.model.rule
 
+import woowacourse.omok.model.board.ColumnStates
+import woowacourse.omok.model.position.Column
+import woowacourse.omok.model.position.Row
 import woowacourse.omok.model.search.Direction
 import woowacourse.omok.model.search.DirectionFirstClearResult
 import woowacourse.omok.model.search.DirectionResult
@@ -9,9 +12,6 @@ import woowacourse.omok.model.stonestate.Black
 import woowacourse.omok.model.stonestate.Clear
 import woowacourse.omok.model.stonestate.StoneState
 import woowacourse.omok.model.stonestate.White
-import woowacourse.omok.model.board.ColumnStates
-import woowacourse.omok.model.position.Column
-import woowacourse.omok.model.position.Row
 
 class LoadMap(private val stoneStates: List<ColumnStates>) : LoadMapInterface {
     override fun loadMap(stone: Stone): Map<Direction, DirectionResult> {
@@ -35,8 +35,8 @@ class LoadMap(private val stoneStates: List<ColumnStates>) : LoadMapInterface {
     }
 
     override fun getCurrentNode(stone: Stone): Node {
-        val row = stone.row.getIndex()
-        val column = stone.column.getIndex()
+        val row = stone.getRowIndex()
+        val column = stone.getColumnIndex()
         return Node(x = column, y = row)
     }
